refactor: isolate tests to 1 process
