+++ rust/libs/telemetry/tests/README.md
@@ -1,0 +1,3 @@
+# Telemetry tests
+
+Sentry's internal state is per-process, hence we test its functionality in integration tests which are spawned as separate processes by cargo.
