+++ rust/libs/telemetry/tests/attach_firezone_id.rs
@@ -1,0 +1,13 @@
+use telemetry::{TESTING, Telemetry};
+
+#[tokio::test]
+async fn set_firezone_id_attaches_user_to_running_session() {
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
+    let mut telemetry = Telemetry::new();
+    telemetry.start("entrypoint", "1.0.0", TESTING);
+
+    Telemetry::set_firezone_id("device-abc".to_owned()).await;
+
+    assert_eq!(Telemetry::current_user().as_deref(), Some("device-abc"));
+}
